+++ src/types/mixinType.ts
@@ -1,0 +1,7 @@
+/**
+ * Defines type that can be used as mixin
+ */
+export interface MixinType<TType = unknown> extends Function 
+{
+    new (): TType;
+}